代码结构修改2.0-2017/9/11
