+++ src/main/java/com/team6/leangoo/controller/ArchiveController.java
@@ -1,5 +1,6 @@
 package com.team6.leangoo.controller;
 
+import com.team6.leangoo.model.User;
 import com.team6.leangoo.service.ArchiveService;
 import com.team6.leangoo.util.AjaxResult;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -13,9 +14,9 @@ import org.springframework.web.bind.annotation.RestController;
  * Time: 11:08
  */
 @RestController
+@RequestMapping("/Archive")
 public class ArchiveController {
     private final ArchiveService archiveService;
-    private Integer userId = 1;
 
     public ArchiveController(ArchiveService archiveService) {
         this.archiveService = archiveService;
@@ -23,9 +24,12 @@ public class ArchiveController {
 
     @RequestMapping(value = "/getArchiveProject",method = RequestMethod.POST)
     public AjaxResult getArchiveProject(){
+        Integer userId = 1;
+        User user = new User();
+        user.setUserId(userId);
         AjaxResult ajaxResult = new AjaxResult();
         try {
-            ajaxResult.setData(archiveService.getArchivePros(userId));
+            ajaxResult.setData(archiveService.getArchivePros(user));
             ajaxResult.seterrcode(0);
             return ajaxResult;
         } catch (Exception e) {
@@ -38,9 +42,12 @@ public class ArchiveController {
 
     @RequestMapping(value = "/getArchiveBoardList",method = RequestMethod.POST)
     public AjaxResult getArchiveBoardList(){
+        Integer userId = 1;
+        User user = new User();
+        user.setUserId(userId);
         AjaxResult ajaxResult = new AjaxResult();
         try {
-            ajaxResult.setData(archiveService.getArchiveBoardList(userId));
+            ajaxResult.setData(archiveService.getArchiveBoardList(user));
             ajaxResult.seterrcode(0);
             return ajaxResult;
         } catch (Exception e) {
