+++ src/main/java/com/team6/leangoo/controller/UserController.java
@@ -1,31 +1,103 @@
 package com.team6.leangoo.controller;
 
-import org.apache.shiro.authz.annotation.RequiresPermissions;
-import org.springframework.stereotype.Controller;
+import com.team6.leangoo.model.User;
+import com.team6.leangoo.service.UserService;
+import com.team6.leangoo.util.AjaxResult;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RestController;
 
-/**
- * Created by AgZou on 2017/8/25.
- * 测试shiro框架用,不含业务逻辑
- */
-@Controller
-@RequestMapping("/userInfo")
+import java.util.HashMap;
+import java.util.Map;
+
+@RestController
+@RequestMapping("/User")
 public class UserController {
-    /**
-     * 用户查询
-     *
-     * @return the string
-     */
-    @RequestMapping("/userDel")
-    @RequiresPermissions("delete")//shiro注解,只有拥有delete权限的用户才能使用(permission表中的name的值)
-    //在GlobalExceptionHandler中配置了假如没权限返回的页面(index)
-    public String userInfo() {
-        return "userInfoDel";
+    private final UserService userService;
+
+    @Autowired
+    public UserController(UserService userService) {
+        this.userService = userService;
     }
 
-    @RequestMapping("/userAdd")
-    @RequiresPermissions("add")//shiro注解,只有拥有add权限的用户才能使用(permission表中的name的值)
-    public String userInfoAdd() {
-        return "userInfoAdd";
+
+    @RequestMapping(value = "/getUserInfoById",method = RequestMethod.POST)
+    public AjaxResult getUserInfoById(){
+        Integer userId = 1;
+        AjaxResult ajaxResult = new AjaxResult();
+        try {
+            User user = new User();
+            user.setUserId(userId);
+            user = userService.getUserInfoById(user);
+            ajaxResult.setData(user);
+            ajaxResult.seterrcode(0);
+            return ajaxResult;
+        } catch (Exception e) {
+            e.printStackTrace();
+            ajaxResult.seterrcode(10);
+            ajaxResult.setinfo("请求失败");
+            return ajaxResult;
+        }
+    }
+
+    @RequestMapping(value = "/changeUserInfo",method = RequestMethod.POST)
+    public AjaxResult changeUserInfo(@RequestBody User userMsg){
+        Integer userId = 1;
+        AjaxResult ajaxResult = new AjaxResult();
+        try {
+            userMsg.setUserId(userId);
+            if (userService.changeUserInfo(userMsg)==1) {
+                ajaxResult.seterrcode(0);
+            } else {
+                ajaxResult.seterrcode(10);
+                ajaxResult.setinfo("操作失败");
+            }
+            User user = new User();
+            user.setUserId(userId);
+            user = userService.getUserInfoById(user);
+            ajaxResult.setData(user);
+            return ajaxResult;
+        } catch (Exception e) {
+            e.printStackTrace();
+            ajaxResult.seterrcode(10);
+            ajaxResult.setinfo("请求失败");
+            return ajaxResult;
+        }
+    }
+
+    @RequestMapping(value = "/changeUserPassword",method = RequestMethod.POST)
+    public AjaxResult changeUserPassword(@RequestBody Map map){
+        Integer userId = 1;
+        AjaxResult ajaxResult = new AjaxResult();
+        try {
+            User user = new User();
+            user.setUserId(userId);
+            user = userService.getUserInfoById(user);
+            String oldPwd = map.get("oldPwd").toString();
+            String newPwd = map.get("newPwd").toString();
+            String newPwdConfirm = map.get("newPwdConfirm").toString();
+            if (user.getUserPassword().equals(oldPwd) && newPwd.equals(newPwdConfirm)){
+                user.setUserPassword(newPwd);
+                if(userService.changeUserInfo(user) == 1) ajaxResult.seterrcode(0);
+                else {
+                    ajaxResult.seterrcode(0);
+                    ajaxResult.setinfo("修改失败");
+                }
+            }else {
+                ajaxResult.seterrcode(0);
+                ajaxResult.setinfo("密码错误，请重新输入");
+            }
+            map = new HashMap();
+            map.put("userPassword",user.getUserPassword());
+            ajaxResult.setData(map);
+            return ajaxResult;
+        } catch (Exception e) {
+            e.printStackTrace();
+            ajaxResult.seterrcode(10);
+            ajaxResult.setinfo("请求失败");
+            return ajaxResult;
+        }
     }
 }
